requirements.txt
streamlit
transformers
faiss-cpu
pymupdf
sentence-transformers
numpy
